Refactor semantic_analyzer.c : make 25 lines
