+++ src/manda/parser/Makefile
@@ -15,6 +15,7 @@ SRCS := tokenizer.c\
 		syntax_action.c \
 		syntax_lr_stack.c \
 		semantic_analyzer.c	\
+		semantic_analyzer_utils.c \
 		
 LDLIBS := -lft
